nlp_training/requirements.txt
# Core data
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0

# ML / NLP
scikit-learn>=1.4.0
scipy>=1.11.0
nltk>=3.8.1
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
transformers>=4.41.0,<5
torch>=2.2.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.20
pydantic>=2.0.0

# Utilities
joblib>=1.4.0
requests>=2.32.0
pypdf>=5.4.0

# Optional: Polars for fast pipeline processing
polars>=1.8.0
